+++ app/src/main/java/top/smartable/albums/ui/viewmodels/MainViewModel.kt
@@ -7,10 +7,10 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
-import top.smartable.albums.data.PhotoBridge
+import top.smartable.albums.data.PhotoGate
 import top.smartable.albums.data.PhotoGatewayApi
-import top.smartable.albums.data.PhotoGatewayApi.DEFAULT_SERVER_URL
 import android.util.Log
+import top.smartable.albums.data.SettingsManager
 
 private const val TAG = "MainViewModel"
 
@@ -23,20 +23,19 @@ class MainViewModel : ViewModel() {
     val logs: StateFlow<List<String>> = _logs.asStateFlow()
 
     init {
-        addLog("Сервис запущен")
-        addLog("Ожидание фото")
-        checkConnection()
+        addLog("Сервис запущен \nОжидание фото")
+//        checkConnection()
     }
 
     fun refreshStatus() {
-        checkConnection()
+//        checkConnection()
     }
 
-    fun checkConnection() {
+    fun checkConnection(context: Context) {
         viewModelScope.launch {
             _statusText.value = "Проверка соединения..."
-            val success = PhotoGatewayApi.ping()
-
+            val serverUrl = SettingsManager.getServerUrl(context)
+            val success = PhotoGatewayApi.ping(serverUrl, context)
             if (success) {
                 _statusText.value = "Соединение есть"
                 addLog("Соединение проверено: OK")
@@ -46,6 +45,7 @@ class MainViewModel : ViewModel() {
             }
         }
     }
+
 
     private fun addLog(message: String) {
         val timestamp = java.text.SimpleDateFormat("HH:mm:ss", java.util.Locale.getDefault())
@@ -57,13 +57,16 @@ class MainViewModel : ViewModel() {
 
 
 
-    fun receivePhotos(context: Context) {
+    fun cloudIt(context: Context) {
         Log.d(TAG, "receivePhotos: функция вызвана")
         viewModelScope.launch {
             try {
                 Log.d(TAG, "receivePhotos: корутина запущена")
                 addLog("Начинаем получение...")
-                val success = PhotoBridge.syncFromGateway(DEFAULT_SERVER_URL, context) { message ->
+                val success = PhotoGate.syncFromGateway(
+                    SettingsManager.getServerUrl(context),
+                    context
+                ) { message ->
                     Log.d(TAG, "Получен лог от PhotoBridge: $message")
                     addLog(message)
                 }
@@ -74,5 +77,9 @@ class MainViewModel : ViewModel() {
                 addLog("Исключение: ${e.message}")
             }
         }
+    }
+
+    fun refreshLogs(context: Context) {
+        _logs.value = SettingsManager.getLogs(context)
     }
 }